slight reffactor
