+++ commands/handlers.go
@@ -1,21 +1,17 @@
-package main
+package commands
 
 import (
 	"discord_bot/ISSNow"
 	"discord_bot/gowiki"
-	"flag"
 	"fmt"
 	"github.com/bwmarrin/discordgo"
+	"github.com/joho/godotenv"
 	"log"
-)
-
-// Bot parameters
-var (
-	guild = flag.String("guild", "", "Test guild ID. If not passed - bot registers commands globally")
+	"os"
 )
 
 var (
-	commands = []*discordgo.ApplicationCommand{
+	Commands = []*discordgo.ApplicationCommand{
 		{
 			Name:        "basic-command",
 			Description: "Basic command",
@@ -80,7 +76,7 @@ var (
 		},
 	}
 
-	commandHandlers = map[string]func(s *discordgo.Session, i *discordgo.InteractionCreate){
+	CommandHandlers = map[string]func(s *discordgo.Session, i *discordgo.InteractionCreate){
 		"basic-command": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
 			s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
 				Type: discordgo.InteractionResponseChannelMessageWithSource,
@@ -90,7 +86,13 @@ var (
 			})
 		},
 		"iss-now": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-			data := ISSNow.GetLocation(cfg.ApiKey)
+			err := godotenv.Load(".env")
+
+			if err != nil {
+				log.Fatalf("Error loading .env file")
+			}
+			apikey := os.Getenv("API_KEY")
+			data := ISSNow.GetLocation(apikey)
 
 			response := fmt.Sprintf("latitude: %s\n longitude: %s\n ", data.Latitude, data.Longitude)
 			fmt.Println(data.Location)
@@ -100,7 +102,7 @@ var (
 				response += fmt.Sprintf("%s \n", data.Location)
 			}
 
-			err := s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+			err = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
 				Type: discordgo.InteractionResponseChannelMessageWithSource,
 				Data: &discordgo.InteractionResponseData{
 					Content: response,
@@ -112,26 +114,32 @@ var (
 		},
 		"wiki-search": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
 			responses := map[discordgo.Locale]string{
-				discordgo.Czech: "Ahoj! Toto je lokalizovaná zpráva",
+				discordgo.Czech:     "Vysledok \n",
+				discordgo.EnglishUS: "Result \n",
 			}
 			options := i.ApplicationCommandData().Options
 			optionMap := make(map[string]*discordgo.ApplicationCommandInteractionDataOption, len(options))
 			for _, opt := range options {
 				optionMap[opt.Name] = opt
 			}
-			response := "Hi! This is a localized message with topic: "
-			if option, ok := optionMap["title"]; ok {
-				response += option.StringValue()
+			response := "If you see this, it's error"
+
+			option, ok := optionMap["článek"]
+			if !ok {
+				option, ok = optionMap["title"]
+			}
+			if ok {
+				if r, ok := responses[i.Locale]; ok {
+					response = r
+				}
 				page, err := gowiki.GetPage(option.StringValue(), -1, false, true)
 				if err != nil {
 					log.Fatal(err)
 				}
-
+				if len(page.URL) < 1 {
+					response += "Not found"
+				}
 				response += page.URL
-			}
-
-			if r, ok := responses[i.Locale]; ok {
-				response = r
 			}
 
 			err := s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
@@ -146,7 +154,8 @@ var (
 		},
 		"wiki-get-random": func(s *discordgo.Session, i *discordgo.InteractionCreate) {
 			responses := map[discordgo.Locale]string{
-				discordgo.Czech: "Ahoj! Toto je lokalizovaná zpráva",
+				discordgo.Czech:     "Random wiki stranky\n",
+				discordgo.EnglishUS: "Random wiki pages\n",
 			}
 			options := i.ApplicationCommandData().Options
 			optionMap := make(map[string]*discordgo.ApplicationCommandInteractionDataOption, len(options))
@@ -154,8 +163,17 @@ var (
 				optionMap[opt.Name] = opt
 			}
 
-			response := fmt.Sprintf("Random wiki pages \n")
-			if option, ok := optionMap["count"]; ok {
+			response := "If you see this, it's error"
+
+			option, ok := optionMap["počet"]
+			if !ok {
+				option, ok = optionMap["count"]
+			}
+
+			if ok {
+				if r, ok := responses[i.Locale]; ok {
+					response = r
+				}
 				pages, err := gowiki.GetRandomPages(int(option.IntValue()))
 				if err != nil {
 					fmt.Println(err)
@@ -165,9 +183,6 @@ var (
 				}
 			}
 
-			if r, ok := responses[i.Locale]; ok {
-				response = r
-			}
 			err := s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
 				Type: discordgo.InteractionResponseChannelMessageWithSource,
 				Data: &discordgo.InteractionResponseData{
@@ -180,42 +195,3 @@ var (
 		},
 	}
 )
-
-func RemoveCommands(s *discordgo.Session, registeredCommands []*discordgo.ApplicationCommand) {
-
-	log.Println("Removing commands...")
-	for _, v := range registeredCommands {
-		err := s.ApplicationCommandDelete(s.State.User.ID, *guild, v.ID)
-		if err != nil {
-			log.Panicf("Cannot delete '%v' command: %v", v.Name, err)
-		}
-	}
-
-	log.Println("Deletion complete")
-
-}
-
-func AddCommands(s *discordgo.Session) []*discordgo.ApplicationCommand {
-	s.AddHandler(func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-		if h, ok := commandHandlers[i.ApplicationCommandData().Name]; ok {
-			h(s, i)
-		}
-	})
-	s.AddHandler(func(s *discordgo.Session, r *discordgo.Ready) {
-		log.Printf("Logged in as: %v#%v", s.State.User.Username, s.State.User.Discriminator)
-	})
-
-	log.Println("Adding commands...")
-	registeredCommands := make([]*discordgo.ApplicationCommand, len(commands))
-	for i, v := range commands {
-		cmd, err := s.ApplicationCommandCreate(s.State.User.ID, *guild, v)
-		if err != nil {
-			log.Panicf("Cannot create '%v' command: %v", v.Name, err)
-		}
-		registeredCommands[i] = cmd
-	}
-	log.Println("Adding commands complete")
-
-	return registeredCommands
-
-}
