+++ commands/slash_commands.go
@@ -1,0 +1,50 @@
+package commands
+
+import (
+	"flag"
+	"github.com/bwmarrin/discordgo"
+	"log"
+)
+
+var (
+	guild = flag.String("guild", "", "Test guild ID. If not passed - bot registers commands globally")
+)
+
+func RemoveCommands(s *discordgo.Session, registeredCommands []*discordgo.ApplicationCommand) {
+
+	log.Println("Removing commands...")
+	for _, v := range registeredCommands {
+		err := s.ApplicationCommandDelete(s.State.User.ID, *guild, v.ID)
+		if err != nil {
+			log.Panicf("Cannot delete '%v' command: %v", v.Name, err)
+		}
+	}
+
+	log.Println("Deletion complete")
+
+}
+
+func AddCommands(s *discordgo.Session) []*discordgo.ApplicationCommand {
+	s.AddHandler(func(s *discordgo.Session, i *discordgo.InteractionCreate) {
+		if h, ok := CommandHandlers[i.ApplicationCommandData().Name]; ok {
+			h(s, i)
+		}
+	})
+	s.AddHandler(func(s *discordgo.Session, r *discordgo.Ready) {
+		log.Printf("Logged in as: %v#%v", s.State.User.Username, s.State.User.Discriminator)
+	})
+
+	log.Println("Adding commands...")
+	registeredCommands := make([]*discordgo.ApplicationCommand, len(Commands))
+	for i, v := range Commands {
+		cmd, err := s.ApplicationCommandCreate(s.State.User.ID, *guild, v)
+		if err != nil {
+			log.Panicf("Cannot create '%v' command: %v", v.Name, err)
+		}
+		registeredCommands[i] = cmd
+	}
+	log.Println("Adding commands complete")
+
+	return registeredCommands
+
+}
