+++ moderation/swearWords.txt
@@ -10,6 +10,7 @@ biatch
 bloody
 blowjob
 blow job
+kkt
 bollock
 bollok
 boner
